pyproject.toml
[tool.poetry]
name = "elevenlab-hackathon"
version = "0.1.0"
description = ""
authors = ["Justin Zhao <justinxzhao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.3"
websockets = "^15.0"
deepgram-sdk = "^3.10.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
mistralai = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
